refactor: move the logic behind a same util for view transition api and  nuxt-loading-indicator
